pyscript/requirements.txt
openmeteo_requests==1.1.0
pandas==2.0.3
requests_cache==1.1.1
retry_requests==2.0.0
#psycopg2==2.9.9
psycopg2-binary==2.9.9
python_dotenv==1.0.0
